+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 - **[UPDATE]** Add possibility to have a role attribute on `<Text>`
 - **[UPDATE]** Add the role "alert" on `<Snackbar>` for accessibility
 - **[UPDATE]** Add aria-hidden on the `<Textfield>` clear button
-- [...]
 - **[UPDATE]** `<Badge>` now also accepts numbers as children
 - **[UPDATE]** `Dependencies` update
 - **[UPDATE]** `ACI & ALUMCIGR` remove deprecated folders for CI
+- **[NEW]** `SuccessModal` component
 
 # v0.20.2 (02/01/2019)